airflow-docker/requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
pandas
pyarrow
duckdb
mysql-connector-python
apache-airflow-providers-mysql
